[LayoutNG] Compute table cell margins to zero.
In the bug we check when we are performing the layout overflow
calculation if we have margins and add them if necessary.

Because we were setting them in LayoutNG (and not in Legacy) it meant
that we were contributing to overflow!

This forces margins to compute to zero for table cells.

Bug: 924504
Change-Id: Ia3595e0b432a965e7ed4bd28a8b3cf046fdf3775